[AssemblyBrowser] Massively improve time it takes to open and a couple of fixes
Reuse the typesystem created for decompilation. No need to create a new one just to load
the treeview.

Prefer using AddChildren everywhere rather than Add. This helps not sort the tree every time
we're adding an element.

Don't bother gathering all the types in an assembly. Most likely we will only need to gather
types from just one assembly, so defer that to the namespace node builder.

Fix multiple iteration of a type's members when trying to load a project metadata view.

And fix nested types to appear in the tree.